requirements.txt
anyjson==0.3.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-boto==0.3.9
djangorestframework==3.2.4
djangorestframework-jsonp==1.0.2
PyMySQL==0.7.2